requirements.txt
cffi==0.8.6
characteristic==14.3.0
cryptography==0.7.2
cssselect==0.9.1
Django==1.7.4
djangorestframework==3.0.5
enum34==1.0.4
gunicorn==19.2.1
httpie==0.9.2
ipython==2.4.1
lxml==3.4.1
psycopg2==2.5.4
pyasn1==0.1.7
pyasn1-modules==0.0.5
pycparser==2.10
Pygments==2.0.2
pyOpenSSL==0.14
queuelib==1.2.2
requests==2.5.3
Scrapy==0.24.4
service-identity==14.0.0
setproctitle==1.1.8
six==1.9.0
Twisted==15.0.0
w3lib==1.11.0
yolk==0.4.3
zope.interface==4.1.2
